+++ R/tt_curate_data.R
@@ -4,20 +4,20 @@
 #' a TidyTuesday dataset. See `vignette("curating", package = "tidytuesdayR)`
 #' for more information.
 #'
+#' @inheritParams usethis::edit_file
+#'
 #' @returns The path to the `tt_curation.R` script, invisibly.
 #' @export
 #'
 #' @examples
 #' tt_curate_data()
-tt_curate_data <- function() {
+tt_curate_data <- function(open = rlang::is_interactive()) {
   path <- system.file("templates", "tt_curation.R", package = "tidytuesdayR")
-  if (rlang::is_interactive()) {
+  if (open) {
     # nocov start
-    if (rlang::is_installed("rstudioapi")) {
-      rstudioapi::documentOpen(path)
-    } else {
-      utils::browseURL(path)
-    }
-  } # nocov end
+    rlang::check_installed("usethis", "to open the curation script.")
+    usethis::edit_file(path)
+    # nocov end
+  }
   return(invisible(path))
 }